refactor: aws mock 중복 코드 fixture로 제거
