+++ blog/tests/admin/test_cleanup_orphan_post_images_command.py
@@ -3,23 +3,13 @@
 from datetime import timedelta
 from unittest.mock import patch
 
-import boto3
 import pytest
 from django.core.management import call_command
 from django.utils import timezone
-from moto import mock_aws
 
 from blog.models import PostImage
 
-
-@pytest.fixture(autouse=True)
-def aws_mock():
-    with mock_aws():
-        boto3.client("s3", region_name="ap-northeast-2").create_bucket(
-            Bucket="dailyfunding-images",
-            CreateBucketConfiguration={"LocationConstraint": "ap-northeast-2"},
-        )
-        yield
+pytestmark = pytest.mark.usefixtures("aws_mock")
 
 
 @pytest.mark.django_db
