+++ blog/tests/admin/test_post_delete_service.py
@@ -1,23 +1,13 @@
 # 게시물 삭제 시 PostImage.post = null / updated_at 갱신 테스트
 from datetime import timedelta
 
-import boto3
 import pytest
 from django.utils import timezone
-from moto import mock_aws
 
 from blog.models import PostImage
 from blog.services.post_delete_service import delete_post
 
-
-@pytest.fixture(autouse=True)
-def aws_mock():
-    with mock_aws():
-        boto3.client("s3", region_name="ap-northeast-2").create_bucket(
-            Bucket="dailyfunding-images",
-            CreateBucketConfiguration={"LocationConstraint": "ap-northeast-2"},
-        )
-        yield
+pytestmark = pytest.mark.usefixtures("aws_mock")
 
 
 @pytest.mark.django_db
